Renommage pages current => old et new => current
